+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerMapping.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.mapping;
+package com.interface21.webmvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
