+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/HandlerAdapterRegistry.java
@@ -1,5 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.adapter;
 
+import com.interface21.webmvc.servlet.HandlerAdapter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
