+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/mapping/AnnotationHandlerMapping.java
@@ -2,6 +2,7 @@ package com.interface21.webmvc.servlet.mvc.mapping;
 
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.HandlerMapping;
 import com.interface21.webmvc.servlet.exception.HandlerInitializationException;
 import com.interface21.webmvc.servlet.mvc.ControllerScanner;
 import com.interface21.webmvc.servlet.mvc.HandlerExecution;
